renamed and deleted files
