--- Input_Reader.go
@@ -1,229 +1,0 @@
-package main
-
-import (
-	"bufio"
-	"errors"
-	"log"
-	"os"
-	"runtime"
-	"strings"
-)
-
-// main entrance for reading user input
-func readInput() {
-	log.Print("QDU-Server launched...\n",
-		"Launch services by entering the respetive settings\n\n")
-
-	for {
-		if err := manageSettings(); err != nil {
-			log.Println(err)
-			continue
-		}
-		break
-	}
-
-	go listenCommands()
-}
-
-func listenCommands() {
-	log.Println("\nWrite '/help' for a list of commands")
-	log.Print("-------------------------------------\n\n")
-	for {
-		s, err := readLine()
-		if err != nil {
-			log.Println(err)
-		}
-		checkCommands(s)
-	}
-}
-
-func manageSettings() error {
-
-	log.Print("Do you want to load a setting file \n",
-		"or create and initialize a new one?\n",
-		"Load [0] / Create [1]\n\n")
-	s, err := readLine()
-	if err != nil {
-		return err
-	}
-	if s == "0" {
-		return loadSettingsFile()
-	}
-	if s == "1" {
-		return createSettingsFile()
-	}
-	return errors.New("input was not in a correct format")
-}
-
-func loadSettingsFile() error {
-	log.Print("Please enter the absolute(!) file path to your config (qdn-file)\n")
-
-	s, err := readLine()
-	if err != nil {
-		return err
-	}
-
-	if err = qdnFileToStruct(s, &config); err != nil {
-		return err
-	}
-
-	log.Print("\nSucessfully loaded config\n")
-
-	return nil
-}
-
-func createSettingsFile() error {
-	log.Print("Please enter the path + (name).qdn to your config file\n\n")
-
-	s, err := readLine()
-	if err != nil {
-		return err
-	}
-	path := s
-
-	log.Print("Please enter your domain <[sub.]domain.toplvl>\n",
-		"(this will be used for creating the URL that is sent to the connected client):\n\n")
-	s, err = readLine()
-	if err != nil {
-		return err
-	}
-	config.Domain = s
-
-	log.Print("Please enter your web directory\n",
-		"(this will be used for creating the URL that is sent to the connected client),",
-		"you can leave this empty:\n\n")
-	s, err = readLine()
-	if err != nil {
-		return err
-	}
-	config.DirectoryWeb = s
-
-	log.Print("Please enter the directory (absolute path) where the images should be saved:\n\n")
-	s, err = readLine()
-	if err != nil {
-		return err
-	}
-	config.DirectoryPics = s
-
-	log.Print("Please enter the desired TCP port:\n\n")
-	s, err = readLine()
-	if err != nil {
-		return err
-	}
-	config.PortTCP = ":" + s
-
-	log.Print("Please enter the desired web port on which this web server will run:\n\n")
-	s, err = readLine()
-	if err != nil {
-		return err
-	}
-	config.PortWeb = ":" + s
-
-	log.Print("Please enter the name of your database:\n\n")
-	s, err = readLine()
-	if err != nil {
-		return err
-	}
-	config.DBName = s
-
-	log.Print("Please enter the user (e.g. root, dbUser) connecting to the database\n",
-		"(this setting is optional depending on your DBMS):\n\n")
-	s, err = readLine()
-	if err != nil {
-		return err
-	}
-	config.DBUser = s
-
-	log.Print("Please enter the password for the previously entered user if set:\n",
-		"(this setting is optional depending on your user):\n\n")
-	s, err = readLine()
-	if err != nil {
-		return err
-	}
-	config.DBPw = s
-
-	log.Print("If you want to use TLS/SSL for securing your connection (https),\n",
-		"you may enter your certificates now. Please enter your fullchain file in this format:\n",
-		"[absolute Path]fullchainname.pem e.g.: /etc/letsencrypt/fullchain.pem \n",
-		"(it's save to proceed without entering anything; http will be used):\n\n")
-	s, err = readLine()
-	if err != nil {
-		return err
-	}
-	config.Fullchain = s
-
-	log.Print("Please enter your certificates private key file formatted like:\n",
-		"[absolute Path]privatekeyname.pem e.g.: /etc/letsencrypt/privkey.pem \n",
-		"(don't enter anything if you haven't for the previous setting):\n\n")
-	s, err = readLine()
-	if err != nil {
-		return err
-	}
-	config.Privkey = s
-
-	return structToQDNFile(path, config)
-}
-
-func readLine() (string, error) {
-	const inputDelimiter = '\n'
-	r := bufio.NewReader(os.Stdin)
-	input, err := r.ReadString(inputDelimiter)
-	if err != nil {
-		return "", err
-	}
-
-	// formats input
-	input = strings.Replace(input, "\n", "", -1)
-	input = strings.Trim(input, " ")
-	if runtime.GOOS == "windows" {
-		input = input[:len(input)-1]
-	}
-	input = strings.Replace(input, " ", "", -1)
-
-	return input, nil
-}
-
-func checkCommands(s string) {
-	// TODO: reduce hard coded strings (via export?)
-	switch {
-	case s == "/getPortTCP":
-		if len(config.PortTCP) < 1 {
-			log.Print("\n\n")
-			break
-		}
-		log.Print(config.PortTCP[1:], "\n\n")
-	case s == "/getPortWeb":
-		if len(config.PortWeb) < 1 {
-			log.Print("\n\n")
-			break
-		}
-		log.Print(config.PortWeb[1:], "\n\n")
-	case s == "/getDirWeb":
-		log.Print(config.DirectoryWeb)
-	case s == "/help":
-		log.Print("\n-------------------------------------------------------------------------------",
-			"\n/help \t\t\t\t- returns a list of commands",
-			"\n/getDomain \t\t\t- gets current domain for links",
-			"\n/getDir \t\t\t- gets output directory of pics",
-			"\n/getDirWeb \t\t\t- gets directory for web requests",
-			"\n/getPortTCP \t\t\t- gets the admin defined port",
-			"\n/getPortWeb \t\t\t- gets the web port",
-			"\n/quit \t\t\t\t- exits the application w/ exit status 3",
-			"\n/info \t\t\t\t- returns information about the programm",
-			"\n\n- Be aware to port-forward to your server and to adjust firewall settings",
-			"\n-------------------------------------------------------------------------------\n\n")
-	case s == "/getDir":
-		log.Println(config.DirectoryPics)
-
-	case s == "/getDomain":
-		log.Println(config.Domain)
-	case s == "/quit":
-		os.Exit(3)
-	case s == "/info":
-		log.Print("\n-------------------------------------------------------------------------------",
-			"\nDevelopment Framework: Golang 1.10",
-			"\nVersion: 1.1 (beta)",
-			"\nCreator: Dominik Ochs",
-			"\n-------------------------------------------------------------------------------\n\n")
-	}
-}
