--- examples/Full_Settings_Windows_Example.qdn
@@ -1,13 +1,0 @@
-configuration
-<
-	Domain=qdu.example.com,
-	DirectoryPics=C:\qdu\pics,
-	DirectoryWeb=/ExampleDir,
-	PortTCP=:1337,
-	PortWeb=:80,
-	DBUser=exampleUser,
-	DBPw=examplePw123,
-	DBName=qdu,
-	Fullchain=C:\Users\User\certs1\fullchain.pem,
-	Privkey=C:\certs2\fullchain.pem,
->
