--- examples/Minimal_Settings.qdn
@@ -1,13 +1,0 @@
-configuration
-<
-	Domain=example.com,
-	DirectoryPics=/var/web/qdu/pics,
-	DirectoryWeb=,
-	PortTCP=:1337,
-	PortWeb=:80,
-	DBUser=,
-	DBPw=,
-	DBName=qdu,
-	Fullchain=,
-	Privkey=,
->
