+++ main.go
@@ -1,31 +1,27 @@
 package main
 
 import (
-	"database/sql"
 	"fmt"
 	"log"
 	"net"
-	"os"
+
+	"github.com/4kills/QDU_server/db"
 
 	"github.com/4kills/base64encoding"
-	_ "github.com/go-sql-driver/mysql"
 )
 
 var config configuration
-var db *sql.DB
 var enc base64encoding.Encoder64
 
 func main() {
-
-	// starts goroutine listening for user input
-	readInput()
-
 	// encoder for shorter links
 	enc = base64encoding.New()
 
 	// establishes connection with database
-	initDB()
-	defer db.Close()
+	if err := db.InitDB(); err != nil {
+		log.Fatal(fmt.Errorf("DB connection error: ", err))
+	}
+	log.Println("Database connection established...")
 
 	// starts web-server for http requests
 	go webServer()
@@ -34,8 +30,7 @@ func main() {
 	// listens for tcp connections through specified port and serves (pic upload, token)
 	ln, err := net.Listen("tcp", config.PortTCP)
 	if err != nil {
-		log.Println("Fatal error:\n", err)
-		os.Exit(1)
+		log.Fatal("Fatal error:\n", err)
 	}
 
 	// allows for any number of concurrent tcp connections
@@ -47,25 +42,5 @@ func main() {
 		}
 
 		go handleClient(conn)
-	}
-}
-
-func initDB() {
-	log.Print("DB-connection established... \n\n")
-	var err error
-
-	if config.DBUser == "" {
-		db, err = sql.Open("mysql", fmt.Sprintf("/%s", config.DBName))
-	} else {
-		db, err = sql.Open("mysql", fmt.Sprintf("%s:%s@/%s", config.DBUser, config.DBPw, config.DBName))
-	}
-
-	if err != nil {
-		log.Fatal("DB open error:", err)
-	}
-
-	err = db.Ping()
-	if err != nil {
-		log.Fatal("DB connection error: ping unsuccessful:", err)
 	}
 }
