+++ qdu.sql
